chore(core): Move `AtomicF64` type into `vector-common`
